requirements.txt
aniso8601==4.1.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
pendulum==2.0.4
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.9
pytzdata==2018.7
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
